refactor(core): finish Global Search relocation into app/search.rs
